Crate cross_domain_message_gossip

Structs

BlockId

ChainMsg

Encoded message with sender info if available.

ChannelDetail

Channel detail between src and dst chain.

ChannelUpdate

Channel update message.

GossipWorker

Gossip worker to gossip incoming and outgoing messages to other peers. Also, streams the decoded extrinsics to destination chain tx pool if available.

GossipWorkerBuilder

Gossip worker builder

Message

A cross chain message with encoded data.

Enums

MessageData

A type of cross chain message

Constants

RELAYER_PREFIX

Functions

can_allow_xdm_submission

get_channel_state

Load the channel state of self_chain_id on src_chain_id.

get_xdm_processed_block_number

Returns the maybe last processed block number for given xdm.

set_channel_state

Set the channel state of self_chain_id on src_chain_id.

set_xdm_message_processed_at

Store the given XDM ID as processed at given block.

start_cross_chain_message_listener

xdm_gossip_peers_set_config

Returns the network configuration for cross chain message gossip.

Type Aliases

ChainSink

Unbounded sender to send encoded message to listeners.