Crate cross_domain_message_gossip

Structs

ChainTxPoolMsg

Encoded message with sender info if available.

GossipWorker

Gossip worker to gossip incoming and outgoing messages to other peers. Also, streams the decoded extrinsics to destination chain tx pool if available.

GossipWorkerBuilder

Gossip worker builder

Message

A cross chain message with encoded data.

Functions

start_cross_chain_message_listener

xdm_gossip_peers_set_config

Returns the network configuration for cross chain message gossip.

Type Aliases

ChainTxPoolSink

Unbounded sender to send encoded ext to listeners.